Added Translation Feature on the Back End
Added Translate route and promised function to translate document text.
A URL to the API is created with the language and text and is called in
a pomised function with a callback that parses the body in JSON.
Next, it finds the correct document by ID and updates it in the database.